[code] Added Patch for ommitting the Ctrl-C Handler
OpenOCD catches several system signals from kernel.
This avoids that the openocd below lua reacts to ctrl+c.

The registered handlers just set a flag, which cause a openOCD
server loop to gently terminate. The romloader_jtag bypasses this server
loop, so the flag has no effect.

Due to problems with loading and unloading the library and
accidentally created shims (unload library) this patch was applied.
It is the hope, that this patch will solve this problem.
One more benefit is, that openocd could finally be stopped during
execution by simple Ctrl+C.